requirements.txt
python-telegram-bot==21.10
openai>=1.0.0
python-dotenv==1.1.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
